vectormagazine-backend/requirements.txt
gunicorn==21.2.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Compress==1.14
psycopg2-binary
python-dotenv==1.0.0
bleach==6.1.0
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.1.2